ТЗ7

+++ src/managers/InMemoryHistoryManager.java
@@ -1,6 +1,7 @@
-package controller;
+package managers;
 
 import base.Task;
+import util.Node;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -8,43 +9,45 @@ import java.util.List;
 import java.util.TreeSet;
 
 public class InMemoryHistoryManager implements HistoryManager {
+
+
     private HashMap<Integer, Node<Task>> nodeIndex = new HashMap<>();   //Индекс поиска задач в истории обращений
     private Node<Task> first;   //Первый узел двусвязного списка истории обращений к задачам
     private Node<Task> last;    //Последний узел двусвязного списка истории обращений к задачам
 
     //Добавление нового просмотра задачи в историю обращений (добавление/перенос в конец списка)
     @Override
-    public void add(Task task) {
+    public void add(Task task){
         Node<Task> l = last;    //Временное сохранение последнего элемента
         Node<Task> newNode = null;
 
-        if (task == null) {
-            return;
+        //Если задача уже есть в списке - взять существующий узел
+        if (nodeIndex.containsKey(task.getNum())){
+            newNode = nodeIndex.get(task.getNum());
+            unlink(newNode);    //Выключение узла из списка перед манипуляциями
         } else {
-            //Если задача уже есть в списке - взять существующий узел
-            if (nodeIndex.containsKey(task.getNum())) {
-                newNode = nodeIndex.get(task.getNum());
-                unlink(newNode);    //Выключение узла из списка перед манипуляциями
-            } else {
-                newNode = new Node<>(l, task, null);   //Создание нового узла
-                nodeIndex.put(task.getNum(), newNode);      //Вставка нового узла в индекс поиска
-            }
-
-            //Включение узла в конец списка
-            linkLast(newNode);
+            newNode = new Node<>(l, task, null);   //Создание нового узла
+            nodeIndex.put(task.getNum(), newNode);      //Вставка нового узла в индекс поиска
         }
+
+        //Включение узла в конец списка
+        linkLast(newNode);
     }
+
     //Удаление просмотра из истории
     @Override
     public void remove(int id){
         Node<Task> delNode = nodeIndex.get(id);
-        unlink(delNode);        //Отсоединение узла из списка
-        nodeIndex.remove(id);   //Удаление узла из индекса поиска
+        if (delNode != null) {
+            unlink(delNode);        //Отсоединение узла из списка
+            nodeIndex.remove(id);   //Удаление узла из индекса поиска
+        }
     }
+
     //Получение истории последних просмотров
     @Override
-    public TreeSet<Task> getHistory(){
-        TreeSet<Task> ret = new TreeSet<>();
+    public List<Task> getHistory(){
+        List<Task> ret = new ArrayList<>();
         for (Node<Task> x = first; x != null; x = x.getNext())
             ret.add(x.getItem());
 
@@ -69,7 +72,9 @@ public class InMemoryHistoryManager implements HistoryManager {
             next.setPrev(prev);
             node.setNext(null);
         }
+
     }
+
     //Процедура включения узла в конец списка
     private void linkLast(Node<Task> node) {
         Node<Task> l = last;    //Временное сохранение последнего элемента
@@ -103,7 +108,7 @@ public class InMemoryHistoryManager implements HistoryManager {
     static List<Integer> fromString(String str){
         List<Integer> ret = new ArrayList<>();
         String[] split = str.split(",");
-        for(String num : split)
+        for (String num : split)
             ret.add(Integer.parseInt(num));
         return ret;
     }
